[HTML] Introduce title row with TOC button
For now, one button is introduced: TOC.
When clicked, it opens up a modal with the table of contents.

After experimenting with a number of different options on how to display
the table of contents, it seemed best to me to just always leave it
behind a button and a modal, so that independent of screen size it works
the same.

I hope that labeling the button with "TOC" is intuitive enough for
people to see that they have to click there to find the table of
contents.

We could change this later for really large screens to still show a
table-of-contents in a left-or-right hand column. I'm not sure at the
moment if that will be best even for large-screen sizes at the moment,
as I'd also like to have margin notes to display things such as
footnotes, todos and maybe references/citations.

Let's improve the HTML look-and-feel incrementally. This seems like a
useful step to commit.

commit-id:1b6fbf2e