[HTML] Only show TOC 2 levels deep
I'm not entirely sure if with the table-of-contents always being in a
modal it makes sense to only show the TOC 2 levels deep.
Only showing 2 levels makes the TOC nicer looking and easier to
navigate. But it does show more details...

Would it be useful to have a toggle in the TOC modal to be able to let
the reader change how many levels of TOC they want to see?

commit-id:54461a10