[HTML] Add a "share" button
... to be able to easily share a pointer to the HTML version of the book
on social media.

This uses https://developer.mozilla.org/en-US/docs/Web/API/Web_Share_API
api where available.

This web share api is mostly not implemented on desktop browsers.

On mobile browsers where it is implemented, this (I hope) pops up a
native "share" functionality, that lets the user choose between
whichever app they have installed to share.

When the web share api is not available, it uses a fall-back technique
with a drop-down menu that shows a number of the most popular social
media. For now, that drop-down menu shows 3 options: twitter, mastodon
and linkedin. We could add more if we think that would be useful.

On Safari on desktop, this web share api is implemented, but does not by
default allow to share on the major social networks, so I thought it
would be better on all desktops to just use the fall-back technique of
using a drop-down menu.